move encrypt condition to rewrite module
